Move message model into separate file
